code optimize
