+++ infra/conf/merge/merge.go
@@ -4,61 +4,72 @@ import (
 	"bytes"
 	"encoding/json"
 	"io"
-	"io/ioutil"
-	"net/http"
-	"net/url"
-	"os"
-	"strings"
-	"time"
 
-	"v2ray.com/core/common/buf"
-	"v2ray.com/core/common/cmdarg"
 	"v2ray.com/core/infra/conf/serial"
 )
 
-// ToJSON merges multiple jsons into one.
-// It accepts []string for URLs, files, [][]byte for json contents
-func ToJSON(args interface{}) ([]byte, error) {
-	m, err := ToMap(args)
+// FilesToJSON merges multiple jsons files into one json, accepts remote url, or local file path
+func FilesToJSON(args []string) ([]byte, error) {
+	m, err := FilesToMap(args)
 	if err != nil {
 		return nil, err
 	}
 	return json.Marshal(m)
 }
 
-// ToMap merges multiple jsons into one map.
-// It accepts []string for URLs, files, [][]byte for json contents
-func ToMap(args interface{}) (m map[string]interface{}, err error) {
-	switch v := args.(type) {
-	case cmdarg.Arg:
-		m, err = filesToMap([]string(v))
-	case []string:
-		m, err = filesToMap(v)
-	case [][]byte:
-		m, err = bytesToMap(v)
-	default:
-		return nil, newError("unsupport args type")
-	}
+// BytesToJSON merges multiple json contents into one json.
+func BytesToJSON(args [][]byte) ([]byte, error) {
+	m, err := BytesToMap(args)
 	if err != nil {
 		return nil, err
 	}
-	sortSlicesInMap(m)
-	err = mergeSameTag(m)
+	return json.Marshal(m)
+}
+
+// FilesToMap merges multiple json files into one map, accepts remote url, or local file path
+func FilesToMap(args []string) (m map[string]interface{}, err error) {
+	m, err = loadFiles(args)
 	if err != nil {
 		return nil, err
 	}
-	removeHelperFields(m)
+	err = applyMergeRules(m)
+	if err != nil {
+		return nil, err
+	}
 	return m, nil
 }
 
-func filesToMap(args []string) (map[string]interface{}, error) {
+// BytesToMap merges multiple json contents into one map.
+func BytesToMap(args [][]byte) (m map[string]interface{}, err error) {
+	m, err = loadBytes(args)
+	if err != nil {
+		return nil, err
+	}
+	err = applyMergeRules(m)
+	if err != nil {
+		return nil, err
+	}
+	return m, nil
+}
+
+func applyMergeRules(m map[string]interface{}) error {
+	sortSlicesInMap(m)
+	err := mergeSameTag(m)
+	if err != nil {
+		return err
+	}
+	removeHelperFields(m)
+	return nil
+}
+
+func loadFiles(args []string) (map[string]interface{}, error) {
 	conf := make(map[string]interface{})
 	for _, arg := range args {
 		r, err := loadArg(arg)
 		if err != nil {
 			return nil, err
 		}
-		m, err := readerToMap(r)
+		m, err := decode(r)
 		if err != nil {
 			return nil, err
 		}
@@ -69,11 +80,11 @@ func filesToMap(args []string) (map[string]interface{}, error) {
 	return conf, nil
 }
 
-func bytesToMap(args [][]byte) (map[string]interface{}, error) {
+func loadBytes(args [][]byte) (map[string]interface{}, error) {
 	conf := make(map[string]interface{})
 	for _, arg := range args {
 		r := bytes.NewReader(arg)
-		m, err := readerToMap(r)
+		m, err := decode(r)
 		if err != nil {
 			return nil, err
 		}
@@ -84,65 +95,11 @@ func bytesToMap(args [][]byte) (map[string]interface{}, error) {
 	return conf, nil
 }
 
-func readerToMap(r io.Reader) (map[string]interface{}, error) {
+func decode(r io.Reader) (map[string]interface{}, error) {
 	c := make(map[string]interface{})
 	err := serial.DecodeJSON(r, &c)
 	if err != nil {
 		return nil, err
 	}
 	return c, nil
-}
-
-// loadArg loads one arg, maybe an remote url, or local file path
-func loadArg(arg string) (out io.Reader, err error) {
-	var data []byte
-	switch {
-	case strings.HasPrefix(arg, "http://"), strings.HasPrefix(arg, "https://"):
-		data, err = fetchHTTPContent(arg)
-	case (arg == "stdin:"):
-		data, err = ioutil.ReadAll(os.Stdin)
-	default:
-		data, err = ioutil.ReadFile(arg)
-	}
-	if err != nil {
-		return
-	}
-	out = bytes.NewBuffer(data)
-	return
-}
-
-// fetchHTTPContent dials https for remote content
-func fetchHTTPContent(target string) ([]byte, error) {
-	parsedTarget, err := url.Parse(target)
-	if err != nil {
-		return nil, newError("invalid URL: ", target).Base(err)
-	}
-
-	if s := strings.ToLower(parsedTarget.Scheme); s != "http" && s != "https" {
-		return nil, newError("invalid scheme: ", parsedTarget.Scheme)
-	}
-
-	client := &http.Client{
-		Timeout: 30 * time.Second,
-	}
-	resp, err := client.Do(&http.Request{
-		Method: "GET",
-		URL:    parsedTarget,
-		Close:  true,
-	})
-	if err != nil {
-		return nil, newError("failed to dial to ", target).Base(err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		return nil, newError("unexpected HTTP status code: ", resp.StatusCode)
-	}
-
-	content, err := buf.ReadAllToBytes(resp.Body)
-	if err != nil {
-		return nil, newError("failed to read HTTP response").Base(err)
-	}
-
-	return content, nil
 }
